Rename entities to better show their purpose
