+++ cs/Markdown.Tests/MdTests.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics;
-using Markdown.Convert;
 using NUnit.Framework;
 
 namespace Markdown.Tests;
@@ -11,8 +10,8 @@ public class MdTests
     [TestCase("__bold__", "<strong>bold</strong>")]
     public void Should_RenderSingleTag(string input, string expectedResult)
     {
-        var converter = new MarkdownToHtmlConverter();
-        var result = converter.Convert(input);
+        var converter = new Md();
+        var result = converter.Render(input);
         Assert.That(result, Is.EqualTo(expectedResult));
     }
     
@@ -22,8 +21,8 @@ public class MdTests
     [TestCase("# heading __bold _italics___", "<h1> heading <strong>bold <em>italics</em></strong></h1>")]
     public void Should_RenderNestedTags(string input, string expectedResult)
     {
-        var converter = new MarkdownToHtmlConverter();
-        var result = converter.Convert(input);
+        var converter = new Md();
+        var result = converter.Render(input);
         Assert.That(result, Is.EqualTo(expectedResult));
     }
     
@@ -34,8 +33,8 @@ public class MdTests
     [TestCase("# heading\r\n __bold__", "<h1> heading\r\n</h1> <strong>bold</strong>")]
     public void Should_RenderTextWithWhitespace(string input, string expectedResult)
     {
-        var converter = new MarkdownToHtmlConverter();
-        var result = converter.Convert(input);
+        var converter = new Md();
+        var result = converter.Render(input);
         Assert.That(result, Is.EqualTo(expectedResult));
     }
 
@@ -46,8 +45,8 @@ public class MdTests
     [TestCase("\\a", "\\a")]
     public void ShouldNot_RenderEscapedTags(string input, string expectedResult)
     {
-        var converter = new MarkdownToHtmlConverter();
-        var result = converter.Convert(input);
+        var converter = new Md();
+        var result = converter.Render(input);
         Assert.That(result, Is.EqualTo(expectedResult));
     }
     
@@ -55,8 +54,8 @@ public class MdTests
     [TestCase("__italics\r\n__", "__italics\r\n__")]
     public void ShouldNot_RenderSeparatedTags(string input, string expectedResult)
     {
-        var converter = new MarkdownToHtmlConverter();
-        var result = converter.Convert(input);
+        var converter = new Md();
+        var result = converter.Render(input);
         Assert.That(result, Is.EqualTo(expectedResult));
     }
     
@@ -64,8 +63,8 @@ public class MdTests
     [TestCase("__aaa_bbb__ccc_", "__aaa_bbb__ccc_")]
     public void ShouldNot_RenderIntersectingTags(string input, string expectedResult)
     {
-        var converter = new MarkdownToHtmlConverter();
-        var result = converter.Convert(input);
+        var converter = new Md();
+        var result = converter.Render(input);
         Assert.That(result, Is.EqualTo(expectedResult));
     }
     
@@ -73,8 +72,8 @@ public class MdTests
     [TestCase("123__456__789", "123__456__789")]
     public void ShouldNot_RenderTagsBetweenNumbers(string input, string expectedResult)
     {
-        var converter = new MarkdownToHtmlConverter();
-        var result = converter.Convert(input);
+        var converter = new Md();
+        var result = converter.Render(input);
         Assert.That(result, Is.EqualTo(expectedResult));
     }
 
@@ -85,10 +84,10 @@ public class MdTests
         var sw = new Stopwatch();
         sw.Reset();
 
-        var converter = new MarkdownToHtmlConverter();
+        var converter = new Md();
         GC.Collect();
         sw.Start();
-        converter.Convert(input);
+        converter.Render(input);
         sw.Stop();
         var elapsed = sw.Elapsed.Ticks;
         sw.Reset();
@@ -97,7 +96,7 @@ public class MdTests
 
         GC.Collect();
         sw.Start();
-        converter.Convert(text!);
+        converter.Render(text!);
         sw.Stop();
 
         Assert.That(sw.Elapsed.Ticks / elapsed, Is.LessThanOrEqualTo(10));
