+++ cs/Markdown/Render/ITokenRenderer.cs
@@ -3,7 +3,7 @@ using Markdown.Token;
 
 namespace Markdown.Render;
 
-public interface IRenderer
+public interface ITokenRenderer
 {
     public string Render(TokenTree tree, Dictionary<Tag, Tag> rules);
 }
